+++ pyramid_skosprovider/utils.py
@@ -5,21 +5,68 @@ This module contains a few utility functions.
 
 import re
 
-from pyramid.renderers import JSON
-
-from skosprovider.skos import (
-    Concept,
-    Collection,
-    Label,
-    Note
+from pyramid_skosprovider.renderers import (
+    json_renderer,
+    concept_adapter,
+    collection_adapter,
+    label_adapter,
+    note_adapter,
+    source_adapter
 )
+
+import logging
+log = logging.getLogger(__name__)
+
+
+class QueryBuilder:
+
+    def __init__(self, request, postprocess=False):
+        self.request = request
+        self.postprocess = postprocess
+        self.no_result = False
+        self.mode = self.request.params.get('mode', 'default')
+        self.language = self.request.params.get(
+            'language',
+            self.request.locale_name
+        )
+        self.label = self.request.params.get('label', None)
+
+    def _build_type(self, query):
+        type = self.request.params.get('type', None)
+        if type in ['concept', 'collection']:
+            query['type'] = type
+        return query
+
+    def _build_label(self, query):
+        if self.label not in [None, '*', '']:
+            if self.mode == 'dijitFilteringSelect' and '*' in self.label:
+                self.postprocess = True
+                query['label'] = self.label.replace('*', '')
+            else:
+                query['label'] = self.label
+        return query
+
+    def _build_collection(self, query):
+        coll = self.request.params.get('collection', None)
+        if coll is not None:
+            query['collection'] = {'id': coll, 'depth': 'all'}
+        return query
+
+    def __call__(self):
+        if self.mode == 'dijitFilteringSelect' and self.label == '':
+            self.no_result = True
+        q = {}
+        q = self._build_type(q)
+        q = self._build_label(q)
+        q = self._build_collection(q)
+        return q
 
 
 def parse_range_header(range):
     '''
     Parse a range header as used by the dojo Json Rest store.
 
-    :param str range: The content of the range header to be parsed. 
+    :param str range: The content of the range header to be parsed.
         eg. `items=0-9`
     :returns: A dict with keys start, finish and number or `False` if the
         range is invalid.
@@ -37,91 +84,3 @@ def parse_range_header(range):
         }
     else:
         return False
-
-json_renderer = JSON()
-
-
-def concept_adapter(obj, request):
-    '''
-    Adapter for rendering a :class:`skosprovider.skos.Concept` to json.
-
-    :param skosprovider.skos.Concept obj: The concept to be rendered.
-    :rtype: :class:`dict`
-    '''
-    return {
-        'id': obj.id,
-        'type': 'concept',
-        'uri': obj.uri,
-        'label': obj.label().label,
-        'concept_scheme': {
-            'uri': obj.concept_scheme.uri,
-            'labels': obj.concept_scheme.labels
-        },
-        'labels': obj.labels,
-        'notes': obj.notes,
-        'narrower': obj.narrower,
-        'broader': obj.broader,
-        'related': obj.related,
-        'member_of': obj.member_of,
-        'subordinate_arrays': obj.subordinate_arrays,
-        'matches': obj.matches
-    }
-
-
-def collection_adapter(obj, request):
-    '''
-    Adapter for rendering a :class:`skosprovider.skos.Collection` to json.
-
-    :param skosprovider.skos.Collection obj: The collection to be rendered.
-    :rtype: :class:`dict`
-    '''
-    return {
-        'id': obj.id,
-        'type': 'collection',
-        'uri': obj.uri,
-        'label': obj.label().label,
-        'concept_scheme': {
-            'uri': obj.concept_scheme.uri,
-            'labels': obj.concept_scheme.labels
-        },
-        'labels': obj.labels,
-        'notes': obj.notes,
-        'members': obj.members,
-        'member_of': obj.member_of,
-        'superordinates': obj.superordinates,
-
-    }
-
-
-def label_adapter(obj, request):
-    '''
-    Adapter for rendering a :class:`skosprovider.skos.Label` to json.
-
-    :param skosprovider.skos.Label obj: The label to be rendered.
-    :rtype: :class:`dict`
-    '''
-    return {
-        'label': obj.label,
-        'type': obj.type,
-        'language': obj.language
-    }
-
-
-def note_adapter(obj, request):
-    '''
-    Adapter for rendering a :class:`skosprovider.skos.Note` to json.
-
-    :param skosprovider.skos.Note obj: The note to be rendered.
-    :rtype: :class:`dict`
-    '''
-    return {
-        'note': obj.note,
-        'type': obj.type,
-        'language': obj.language,
-        'markup': obj.markup
-    }
-
-json_renderer.add_adapter(Concept, concept_adapter)
-json_renderer.add_adapter(Collection, collection_adapter)
-json_renderer.add_adapter(Label, label_adapter)
-json_renderer.add_adapter(Note, note_adapter)